설명 문서 정리

+++ DirectX12/D3D12_1/Project60/Scene.h
@@ -65,7 +65,7 @@ public:
 
 	void ReleaseUploadBuffers();
 
-	CGameObject* PickObjectPointedByCursor(int xClient, int yClient, CCamera* pCamera);
+	CGameObject* PickObjectPointedByCursor(float xClient, float yClient, CCamera* pCamera);
 
 
 	CPlayer* m_pPlayer = NULL;
